calculating information criteria from a coxtp object

This function is to calculate information criteria from a coxtp object to select the penalization tuning paramter.

IC(fit, IC.prox, ...)

Arguments

fit

model from coxtp.

IC.prox

when calculating information criteria, there might be numerical issue(second order derivative, Hessian matrix approximate is singular). In such cases warnings will be given. if true, we modified the diagonal of hessian matrix a bit, which may lead to bias issues. Default is FALSE.

Value

model.mAIC

an object with S3 class "coxtp" using mAIC to select the tunning parameter.

model.TIC

an object with S3 class "coxtp" using TIC to select the tunning parameter.

model.GIC

an object with S3 class "coxtp" using GIC to select the tunning parameter.

mAIC

a sequence of mAIC values for the different tuning parameters lambda from "coxtp".

TIC

a sequence of TIC values for the different tuning parameters lambda from "coxtp".

GIC

a sequence of GIC values for the different tuning parameters lambda from "coxtp".

Details

In order to select the proper smoothing parameter, we utilize the idea of information criteria. We provide four different information criteria to select the optimal smoothing parameter \(\lambda\). Generally, mAIC, TIC and GIC selects similar parameters and the difference of resulting estimations are barely noticable. See details in Lingfeng Luo et al. (2022)

Examples

data(ExampleData)
z <- ExampleData$x
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtp(event = event, z = z, time = time)
#> Iter 1: Obj fun = -3.2986480; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.2920862; Stopping crit = 2.1176619e-02;
#> Iter 3: Obj fun = -3.2920355; Stopping crit = 1.6361107e-04;
#> Iter 4: Obj fun = -3.2920353; Stopping crit = 3.7757230e-07;
#> Iter 5: Obj fun = -3.2920353; Stopping crit = 5.0225062e-12;
#> Iter 1: Obj fun = -3.3017443; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.2954100; Stopping crit = 2.0664261e-02;
#> Iter 3: Obj fun = -3.2953323; Stopping crit = 2.5346890e-04;
#> Iter 4: Obj fun = -3.2953321; Stopping crit = 4.6097119e-07;
#> Iter 5: Obj fun = -3.2953321; Stopping crit = 1.7713469e-12;
#> Iter 1: Obj fun = -3.3094946; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.3042222; Stopping crit = 1.7708812e-02;
#> Iter 3: Obj fun = -3.3041817; Stopping crit = 1.3597425e-04;
#> Iter 4: Obj fun = -3.3041817; Stopping crit = 6.0903559e-08;
#> Iter 5: Obj fun = -3.3041817; Stopping crit = 1.4914076e-15;
IC  <- IC(fit)
#> Error in IC(fit): object 'fit.smoothspline' not found