Cox Non-proportional Hazards model with penalization: Main function

Cox Non-proportional Hazards model with penalization: Main function

coxtp(
  event,
  z,
  time,
  strata = c(),
  spline = "Smooth-spline",
  nsplines = 8,
  ties = "Breslow",
  tol = 1e-09,
  iter.max = 20L,
  method = "Newton",
  lambda = 1e+08,
  btr = "static",
  tau = 0.5,
  stop = "ratch",
  parallel = FALSE,
  threads = 1L,
  degree = 3L,
  TIC = FALSE,
  TIC_prox = FALSE,
  lambda_spline = 0,
  ord = 4,
  fixedstep = FALSE,
  ICLastOnly = TRUE,
  penalizestop = FALSE
)

Arguments

event

event vector, should be a vector containing 0 or 1

z

Covariate matrix

time

Time vector, should be a vector with non-negative numeric value

strata

stratification group defined in the data. If there exist stratification group, please enter as vector.

spline

The spline term for Penalized Newton's Method(Add section Number related to the paper). Default setting is spline="Smooth-spline"

nsplines

Number of base functions in the B-splines, default is 8.

ties

Ways to deal with ties, default is ties="Breslow":

tol

Convergence threshold. The default threshold is set as tol=1e-6

iter.max

Maximum Iteration number, default is iter.max=20L

method

Selecting Method used, default is method="Newton"

lambda

Parameter for Proximal Newton's Method. Default is lambda=1e8

btr

Backtracking line search approach, default is btr="static":

tau

(Alpha or beta?) in the Newton's Method, Default is tau=0.5. Used to control for step size.

stop

Stopping rule, default is stop="ratch":

parallel

Parallel computation, Default is parallel=FALSE

threads

Parallel computation parameter(number of cores)Default is threads=1L

degree

Degree of smoothing spline. Default setting is degree=3L.

TIC

TIC_prox

When calculating information criteria, there might be numerical issue(second order derivative, Hessian matrix approximate is singular), thus we proposed to add a small term to the diagonal. Default TIC_prox = FALSE

lambda_spline

Smoothing parameter lambda. Default is lambda_spline = 0 which refers to Newton's Method without penalization.

ord

Specify which derivative to penalize. Default setting is ord=4.

fixedstep

There might be times when the stopping criteria not working, thus, the number of steps could be set manually. Default value is fixedstep = FALSE, if it is true, will stop by iter.max

ICLastOnly

Only calculate the last information criteria if is TRUE. Default is ICLastOnly=FALSE

penalizestop