study to Understand Prognoses Preferences Outcomes and Risks of Treatment (SUPPORT)

The SUPPORT dataset tracks four response variables: hospital death, severe functional disability, hospital costs, and time until death and death itself. The patients are followed for up to 5.56 years. Data included only tracks follow-up time and death.

support

Format

A dataframe with 9104 observations and 34 variables after imputation and the removal of response variables like hospital charges, patient ratio of costs to charges and micro-costs. Ordinal variables, namely functional disability and income, were also removed. Finally, Surrogate activities of daily living were removed due to sparsity. There were 6 other model scores in the data-set and they were removed; only aps and sps were kept.

Age

Stores a double representing age.

death

Death at any time up to NDI date: 31DEC94.

sex

0=female, 1=male.

slos

Days from study entry to discharge.

d.time

days of follow-up.

dzgroup

Each level of dzgroup: ARF/MOSF w/Sepsis, COPD, CHF, Cirrhosis, Coma, Colon Cancer, Lung Cancer, MOSF with malignancy.

dzclass

ARF/MOSF, COPD/CHF/Cirrhosis, Coma and cancer disease classes.

num.co

the number of comorbidities.

edu

years of education of patient.

scoma

The SUPPORT coma score based on Glasgow D3.

avtisst

Average TISS, days 3-25.

race

Indicates race. White, Black, Asian, Hispanic or other.

hday

Day in Hospital at Study Admit

diabetes

Diabetes (Com 27-28, Dx 73)

dementia

Dementia (Comorbidity 6)

ca

Cancer State

meanbp

Mean Arterial Blood Pressure Day 3.

wblc

White blood cell count on day 3.

hrt

Heart rate day 3.

resp

Respiration Rate day 3.

temp

Temperature, in Celsius, on day 3.

pafi

PaO2/(0.01*FiO2) Day 3.

alb

Serum albumin day 3.

bili

Bilirubin Day 3.

crea

Serum creatinine day 3.

sod

Serum sodium day 3.

ph

Serum pH (in arteries) day 3.

glucose

Serum glucose day 3.

bun

BUN day 3.

urine

urine output day 3.

adlp

ADL patient day 3.

adlsc

Imputed ADL calibrated to surrogate, if a surrogate was used for a follow up.

sps

SUPPORT physiology score

aps

Apache III physiology score

Source

Available at the following website: https://biostat.app.vumc.org/wiki/Main/SupportDesc. note: must unzip and process this data before use. #' @examples data("support")

Details

Some of the original data was missing. Before imputation, there were a total of 9105 individuals and 47 variables. Of those variables, a few were removed before imputation. We removed three response variables: hospital charges, patient ratio of costs to charge,s and patient micro-costs. Next, we removed hospital death as it was directly informative of our event of interest, namely death. We also removed functional disability and income as they are ordinal covariates. Finally, we removed 8 covariates related to the results of previous findings: we removed SUPPORT day 3 physiology score (sps), APACHE III day 3 physiology score (aps), SUPPORT model 2-month survival estimate, SUPPORT model 6-month survival estimate, Physician's 2-month survival estimate for pt., Physician's 6-month survival estimate for pt., Patient had Do Not Resuscitate (DNR) order, and Day of DNR order (<0 if before study). Of these, sps and aps were added on after imputation, as they were missing only 1 observation. First we imputed manually using the normal values for physiological measures recommended by Knaus et al. (1995). Next, we imputed a single dataset using mice with default settings. After imputation, we noted that the covariate for surrogate activities of daily living was not imputed. This is due to collinearity between the other two covariates for activities of daily living. Therefore, surrogate activities of daily living was removed.