testing the significance of the covariates from a coxtv or coxtp object

Testing the significance of the covariates from a coxtv or coxtp object using a Wald test statistic. The null hypothesis \(H_0: \beta(t) = 0\) for any \(t\), where \(t\) denotes the event time.

tvef.zero(fit, parm)

Arguments

fit

fitted coxtv or coxtp model.

parm

the names of parameters to be tested.

Value

tvef.zero produces a matrix. Each row corresponds to a covariate from the fitted object. The three columns give the test statistic, degrees of freedom and P-value.

See also

Examples

data(ExampleData)
z <- ExampleData$z
time <- ExampleData$time
event <- ExampleData$event
fit <- coxtv(event = event, z = z, time = time)
#> Iter 1: Obj fun = -3.2982771; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.2916285; Stopping crit = 2.1424865e-02;
#> Iter 3: Obj fun = -3.2916034; Stopping crit = 8.0884492e-05;
#> Iter 4: Obj fun = -3.2916034; Stopping crit = 1.8232153e-09;
#> Algorithm converged after 4 iterations!
tvef.ph(fit)
#>         chisq df            p
#> X1   7.013482  7 4.274774e-01
#> X2 147.312907  7 1.488153e-28