fit a Cox non-proportional hazards model with p-spline or smoothing-spline, penalization tuning parameter can be chosen by information criteria or cross-validation.

Fit a Cox non-proportional hazards model via penalized maximum likelihood.

coxtp(
  event,
  z,
  time,
  strata = NULL,
  penalty = "Smooth-spline",
  nsplines = 8,
  lambda = c(0.1, 1, 10),
  degree = 3L,
  knots = NULL,
  ties = "Breslow",
  tol = 1e-09,
  iter.max = 20L,
  method = "ProxN",
  gamma = 1e+08,
  btr = "dynamic",
  tau = 0.5,
  stop = "ratch",
  parallel = FALSE,
  threads = 1L,
  fixedstep = FALSE,
  InfoCrit = FALSE,
  ...
)

Arguments

event

failure events response variable of length nobs, where nobs denotes the number of observations. It should be a vector containing 0 or 1

z

input covariate matrix, of dimension nobs x nvars; each row is an observation vector.

time

observed event time, should be a vector with non-negative numeric values.

strata

stratification group defined in the data used for the stratified model. If there exists a stratification group, please enter it as a vector. By default, a non-stratified model would be implemented.

penalty

a character string specifying the spline term for Penalized Newton's Method. This term is added to the log-partial likelihood as the new objective function to control the smoothness of the time-varying covariates. Default is P-spline. Three options are P-spline, Smooth-spline and NULL. If NULL, the method will be the same as coxtv and lambda will be set as 0.

P-spline stands for Penalized B-spline. It combines the B-spline basis with a discrete quadratic penalty on the difference of basis coefficients between adjacent knots. When lambda goes to infinity, the time-varying effects are reduced to be constant.

Smooth-spline refers to the Smoothing-spline, the derivative-based penalties combined with B-splines. See degree for different choices. When degree=3, we use the cubic B-spline penalizing the second-order derivative, which reduces the time-varying effect to a linear term when lambda goes to infinity. When degree=2, we use the quadratic B-spline penalizing first-order derivative, which reduces the time-varying effect to a constant when lambda goes to infinity. See Wood (2016) for details.

If P-spline or Smooth-spline, then lambda is initialized as (0.1, 1, 10). Users can modify lambda. See details in lambda.

nsplines

number of basis functions in the B-splines to span the time-varying effects, the default value is 8. We use the r function splines::bs to generate the B-splines.

lambda

a user-specified lambda sequence as the penalization coefficients in front of the spline term specified by spline. This is the tuning parameter for penalization. Users can use IC to select the best tuning parameter based on the information criteria. Users can specify larger values when the absolute values of the estimated time-varying effects are too large. Default is 0, which refers to Newton's Method without penalization.

degree

degree of the piecewise polynomial for generating the B-spline basis functions---default is 3 for cubic splines. degree = 2 results in the quadratic B-spline basis functions.

If penalty is P-spline or NULL, degree's default value is 3.

If penalty is Smooth-spline, degree's default value is 2.

knots

the internal knot locations (breakpoints) that define the B-splines. The number of the internal knots should be nsplines-degree-1. If NULL, the locations of knots are chosen to include an equal number of events within each time interval. This choice leads to more stable results in most cases. Users can specify the internal knot locations by themselves.

ties

a character string specifying the method for tie handling. If there are no tied death times, the methods are equivalent. By default "Breslow" uses the Breslow approximation, which can be faster when many ties occur.

tol

convergence threshold for Newton's method. The algorithm continues until the method selected using stop converges. The default value is 1e-6.

iter.max

maximum Iteration number if the stopping criteria specified by stop is not satisfied. Default value is 20.

method

a character string specifying whether to use Newton's method or Proximal Newton's method. If "Newton" then exact hessian is used, while default method "ProxN" implementing the proximal method which can be faster and more stable when there exists ill-conditioned second-order information of the log-partial likelihood. See details in Wu et al. (2022).

gamma

parameter for Proximal Newton's Method "ProxN". The default value is 1e8.

btr

a character string specifying the backtracking line-search approach. "dynamic" is a typical way to perform backtracking line-search. See details in Convex Optimization by Boyd and Vandenberghe (2009). "static" limits Newton's increment and can achieve more stable results in some extreme cases, such as ill-conditioned second-order information of the log-partial likelihood, which usually occurs when some predictors are categorical with low frequency for some categories. Users should be careful with static as this may lead to under-fitting.

tau

a scalar in (0,1) used to control the step size inside the backtracking line-search. The default value is 0.5.

stop

a character string specifying the stopping rule to determine convergence. Use \(loglik(m)\) to denote the log-partial likelihood at iteration step m. "incre" means we stop the algorithm when Newton's increment is less than the tol. "relch" means we stop the algorithm when the \(loglik(m)\) divided by the \(loglik(0)\) is less than the tol. "ratch" means we stop the algorithm when \((loglik(m)-loglik(m-1))/(loglik(m)-loglik(0))\) is less than the tol. "all" means we stop the algorithm when all the stopping rules "incre", "relch" and "ratch" is met. Default value is ratch. If the maximum iteration steps iter.max is achieved, the algorithm stops before the stopping rule is met.

parallel

if TRUE, then the parallel computation is enabled. The number of threads in use is determined by threads.

threads

an integer indicating the number of threads to be used for parallel computation. Default is 2. If parallel is false, then the value of threads has no effect.

fixedstep

if TRUE, the algorithm will be forced to run iter.max steps regardless of the stopping criterion specified.

Value

A list of objects with S3 class "coxtp". The length is the same as that of lambda; each represents the model output with each value of the tuning parameter lambda.

call

the call that produced this object.

beta

the estimated time varying coefficient for each predictor at each unique time. It is a matrix of dimension len_unique_t x nvars, where len_unique_t is the length of unique follow-up time. Each row represents the coefficients at the corresponding input observation time.

bases

the basis matrix used in model fitting. If ties="None", the dimension is nvars x nsplines; if ties="Breslow", the dimension is len_unique_t x nsplines. The matrix is constructed using bs::splines function.

ctrl.pts

estimated coefficient of the basis matrix of dimension nvars x nsplines. Each row represents a covariate's coefficient on the nsplines dimensional basis functions.

Hessian

the Hessian matrix of the log-partial likelihood, of which the dimension is nsplines * nvars x nsplines * nvars.

internal.knots

the internal knot locations of the basis functions. The locations of knots are chosen to include an equal number of events within each time interval.

nobs

number of observations.

spline

the spline type user specified.

theta.list

a list of ctrl.pts of length m, contains the updated ctrl.pts after each algorithm iteration.

VarianceMatrix

the variance matrix of the estimated coefficients of the basis matrix, which is the inverse of the negative Hessian matrix.

Details

The sequence of models implied by lambda.spline is fit by Newton's method (Proximal Newton's method). The objective function is $$loglik - P_{\lambda}$$, where \(P_{\lambda}\) can be P-spline or Smooth-spline. The \(\lambda\) is the tuning parameter \(\lambda\). Users can define the initial sequence. ICprovides different information criteria to choose the tuning parameter \(\lambda\). cv.coxtp uses the cross validation to choose the tuning parameter.

References

Gray, Robert J. (1992) Flexible methods for analyzing survival data using splines, with applications to breast cancer prognosis. Journal of the American Statistical Association, 87: 942-951.

Gray, Robert J. (1994) Spline-based tests in survival analysis. Biometrics, 50: 640-652.

Lingfeng Luo, Kevin He, Wenbo Wu and Jeremy M.G. Taylor. (2022) Using information criteria to select smoothing parameters when analyzing survival data with time-varying coefficient hazard models.

Wenbo Wu, Jeremy M.G. Taylor, Andrew F Brouwer, Lingfeng Luo, Jian Kang, Hui Jiang and Kevin He. (2022) Scalable proximal methods for cause-specific hazard modeling with time-varying coefficients Lifetime Data Analysis, 28(2): 194-218.

Wood, Simon N. (2017) P-splines with derivative based penalties and tensor product smoothing of unevenly distributed data. Statistics and Computing, 27(4): 985-989.

Perperoglou, Aris, Saskia le Cessie, and Hans C. van Houwelingen. (2006) A fast routine for fitting Cox models with time varying effects of the covariates. Computer Methods and Programs in Biomedicine, 81(2): 154-161.

See also

Examples

data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event

lambda  = c(0,1)
fit   <- coxtp(event = event, z = z, time = time, lambda=lambda)
#> Iter 1: Obj fun = -3.2982771; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.2916285; Stopping crit = 2.1424865e-02;
#> Iter 3: Obj fun = -3.2916034; Stopping crit = 8.0884492e-05;
#> Iter 4: Obj fun = -3.2916034; Stopping crit = 1.8232153e-09;
#> Iter 5: Obj fun = -3.2916034; Stopping crit = 2.2895367e-14;
#> Iter 1: Obj fun = -3.3017443; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.2954100; Stopping crit = 2.0664261e-02;
#> Iter 3: Obj fun = -3.2953323; Stopping crit = 2.5346890e-04;
#> Iter 4: Obj fun = -3.2953321; Stopping crit = 4.6097119e-07;
#> Iter 5: Obj fun = -3.2953321; Stopping crit = 1.7713469e-12;