fit a Cox non-proportional hazards model with p-spline or smoothing-spline where penalization tuning parameter is provided by cross validation.

Fit a Cox non-proportional hazards model via penalized maximum likelihood.

cv.coxtp(
  event,
  z,
  time,
  strata = NULL,
  lambda = c(0.1, 1, 10),
  nfolds = 5,
  foldid = NULL,
  penalty = "Smooth-spline",
  nsplines = 8,
  ties = "Breslow",
  tol = 1e-09,
  iter.max = 20L,
  method = "ProxN",
  gamma = 1e+08,
  btr = "dynamic",
  tau = 0.5,
  stop = "ratch",
  parallel = FALSE,
  threads = 1L,
  degree = 3L,
  fixedstep = FALSE
)

Arguments

event

failure events response variable of length nobs, where nobs denotes the number of observations. It should be a vector containing 0 or 1

z

input covariate matrix, of dimension nobs x nvars; each row is an observation vector.

time

observed event time, should be a vector with non-negative numeric values.

strata

stratification group defined in the data used for the stratified model. If there exists a stratification group, please enter it as a vector. By default, a non-stratified model would be implemented.

lambda

a user specified sequence as the penalization coefficients in front of the spline term specified by spline. This is the tuning parameter for penalization. Users can use IC to select the best tuning parameter based on the information criteria. Users can specify for larger values when the estimated time-varying effects are too high. Default is 0 which refers to Newton's Method without penalization.

nfolds

number of folds for cross-validation - default is 5. The smallest value allowable is nfolds=3.

foldid

an optional vector of values between 1 and nfold identifying what fold each observation is in. If supplied, nfolds can be missing.

penalty

a character string specifying the spline term for Penalized Newton's Method. This term is added to the log-partial likelihood as the new objective function to control the smoothness of the time-varying covariates. Default is P-spline. Three options are P-spline, Smooth-spline and NULL. If NULL, the method will be the same as coxtv and lambda will be set as 0.

P-spline stands for Penalized B-spline. It combines the B-spline basis with a discrete quadratic penalty on the difference of basis coefficients between adjacent knots. When lambda goes to infinity, the time-varying effects are reduced to be constant.

Smooth-spline refers to the Smoothing-spline, the derivative-based penalties combined with B-splines. See degree for different choices. When degree=3, we use the cubic B-spline penalizing the second-order derivative, which reduces the time-varying effect to a linear term when lambda goes to infinity. When degree=2, we use the quadratic B-spline penalizing first-order derivative, which reduces the time-varying effect to a constant when lambda goes to infinity. See Wood (2016) for details.

If P-spline or Smooth-spline, then lambda is initialized as (0.1, 1, 10). Users can modify lambda. See details in lambda.

nsplines

number of basis functions in the B-splines to span the time-varying effects, default value is 8. We use the r function splines::bs to generate the B-splines.

ties

a character string specifying the method for tie handling. If there are no tied death times, the methods are equivalent. By default "Breslow" uses the Breslow approximation, which can be faster when many ties occur.

tol

convergence threshold for Newton's method. The algorithm continues until the method selected using stop converges. The default value is 1e-6.

iter.max

maximum Iteration number if the stopping criteria specified by stop is not satisfied. Default value is 20.

method

a character string specifying whether to use Newton's method or Proximal Newton's method. If "Newton" then exact hessian is used, while default method "ProxN" implementing the proximal method which can be faster and more stable when there exists ill-conditioned second-order information of the log-partial likelihood. See details in Wu et al. (2022).

gamma

parameter for Proximal Newton's Method "ProxN". The default value is 1e8.

btr

a character string specifying the backtracking line-search approach. "dynamic" is a typical way to perform backtracking line-search. See details in Convex Optimization by Boyd and Vandenberghe (2009). "static" limits Newton's increment and can achieve more stable results in some extreme cases, such as ill-conditioned second-order information of the log-partial likelihood, which usually occurs when some predictors are categorical with low frequency for some categories. Users should be careful with static as this may lead to under-fitting.

tau

a scalar in (0,1) used to control the step size inside the backtracking line-search. The default value is 0.5.

stop

a character string specifying the stopping rule to determine convergence. Use \(loglik(m)\) to denote the log-partial likelihood at iteration step m. "incre" means we stop the algorithm when Newton's increment is less than the tol. "relch" means we stop the algorithm when the \(loglik(m)\) divided by the \(loglik(0)\) is less than the tol. "ratch" means we stop the algorithm when \((loglik(m)-loglik(m-1))/(loglik(m)-loglik(0))\) is less than the tol. "all" means we stop the algorithm when all the stopping rules "incre", "relch" and "ratch" is met. Default value is ratch. If the maximum iteration steps iter.max is achieved, the algorithm stops before the stopping rule is met.

parallel

if TRUE, then the parallel computation is enabled. The number of threads in use is determined by threads.

threads

an integer indicating the number of threads to be used for parallel computation. Default is 2. If parallel is false, then the value of threads has no effect.

degree

degree of the piecewise polynomial for generating the B-spline basis functions---default is 3 for cubic splines. degree = 2 results in the quadratic B-spline basis functions.

If penalty is P-spline or NULL, degree's default value is 3.

If penalty is Smooth-spline, degree's default value is 2.

fixedstep

if TRUE, the algorithm will be forced to run iter.max steps regardless of the stopping criterion specified.

knots

the internal knot locations (breakpoints) that define the B-splines. The number of the internal knots should be nsplines-degree-1. If NULL, the locations of knots are chosen to include an equal number of events within each time interval. This choice leads to more stable results in most cases. Users can specify the internal knot locations by themselves.

Value

an object of class "cv.coxtp" is returned, which is a list with the ingredients of the cross-validation fit.

model.cv

a "coxtp" object with tuning parameter chosen based on cross validation

lambda

the values of lambda used in the fits.

cve

the mean cross-validated error - a vector of length(lambda). For the k-th testing fold (k = 1,...,nfolds), we take the remaining folds as the training folds. Based on the model trained on the training folds, we calculate the log-partial likelihood on all the folds \(loglik0\) and training folds \(loglik1\). The CVE is equal to \(-2*(loglik0 - loglik1)\). This approach avoids the construction of a partial likelihood on the test set so that the risk set is always sufficiently large.

lambda.min

the value of lambda that gives minimum cve.

Details

The function runs coxtp length of lambda x nfolds times; each is to compute the fit with each of the folds omitted.

References

Gray, Robert J. (1992) Flexible methods for analyzing survival data using splines, with applications to breast cancer prognosis. Journal of the American Statistical Association, 87: 942-951.

Gray, Robert J. (1994) Spline-based tests in survival analysis. Biometrics, 50: 640-652.

Lingfeng Luo, Kevin He, Wenbo Wu and Jeremy M.G. Taylor. (2022) Using information criteria to select smoothing parameters when analyzing survival data with time-varying coefficient hazard models.

Verweij, Pierre JM, and Hans C. Van Houwelingen. (1993) Cross‐validation in survival analysis. Statistics in Medicine, 12(24): 2305-2314.

Wenbo Wu, Jeremy M.G. Taylor, Andrew F Brouwer, Lingfeng Luo, Jian Kang, Hui Jiang and Kevin He. (2022) Scalable proximal methods for cause-specific hazard modeling with time-varying coefficients Lifetime Data Analysis, 28(2): 194-218.

Wood, Simon N. (2017) P-splines with derivative based penalties and tensor product smoothing of unevenly distributed data. Statistics and Computing, 27(4): 985-989.

Perperoglou, Aris, Saskia le Cessie, and Hans C. van Houwelingen. (2006) A fast routine for fitting Cox models with time varying effects of the covariates. Computer Methods and Programs in Biomedicine, 81(2): 154-161.

Examples

data(ExampleData)
z <- ExampleData$x
time  <- ExampleData$time
event <- ExampleData$event
lambda  = c(0.1, 1)
fit  <- cv.coxtp(event = event, z = z, time = time, lambda=lambda, nfolds = 5)
#> Iter 1: Obj fun = -3.1855946; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1784225; Stopping crit = 2.3334549e-02;
#> Iter 3: Obj fun = -3.1783394; Stopping crit = 2.7014970e-04;
#> Iter 4: Obj fun = -3.1783391; Stopping crit = 1.0032555e-06;
#> Iter 5: Obj fun = -3.1783391; Stopping crit = 2.7977920e-11;
#> Iter 1: Obj fun = -3.1895056; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1829265; Stopping crit = 2.1723742e-02;
#> Iter 3: Obj fun = -3.1828395; Stopping crit = 2.8707554e-04;
#> Iter 4: Obj fun = -3.1828394; Stopping crit = 4.8807358e-07;
#> Iter 5: Obj fun = -3.1828394; Stopping crit = 1.6902138e-12;
#> Iter 1: Obj fun = -3.1578498; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1505975; Stopping crit = 2.1889158e-02;
#> Iter 3: Obj fun = -3.1505308; Stopping crit = 2.0114281e-04;
#> Iter 4: Obj fun = -3.1505307; Stopping crit = 5.2713242e-07;
#> Iter 5: Obj fun = -3.1505307; Stopping crit = 7.2230805e-12;
#> Iter 1: Obj fun = -3.1626625; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1557710; Stopping crit = 2.1129979e-02;
#> Iter 3: Obj fun = -3.1556827; Stopping crit = 2.7064361e-04;
#> Iter 4: Obj fun = -3.1556825; Stopping crit = 4.7615624e-07;
#> Iter 5: Obj fun = -3.1556825; Stopping crit = 1.6620882e-12;
#> Iter 1: Obj fun = -3.1981592; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1924811; Stopping crit = 1.9032636e-02;
#> Iter 3: Obj fun = -3.1924417; Stopping crit = 1.3213820e-04;
#> Iter 4: Obj fun = -3.1924416; Stopping crit = 3.0834610e-07;
#> Iter 5: Obj fun = -3.1924416; Stopping crit = 3.2237966e-12;
#> Iter 1: Obj fun = -3.2012595; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1957760; Stopping crit = 1.8585366e-02;
#> Iter 3: Obj fun = -3.1957198; Stopping crit = 1.9062928e-04;
#> Iter 4: Obj fun = -3.1957197; Stopping crit = 3.0143171e-07;
#> Iter 5: Obj fun = -3.1957197; Stopping crit = 8.6832603e-13;
#> Iter 1: Obj fun = -3.1871817; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1808281; Stopping crit = 2.0750560e-02;
#> Iter 3: Obj fun = -3.1807870; Stopping crit = 1.3435269e-04;
#> Iter 4: Obj fun = -3.1807869; Stopping crit = 2.5143548e-07;
#> Iter 5: Obj fun = -3.1807869; Stopping crit = 2.6625299e-12;
#> Iter 1: Obj fun = -3.1910907; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1848813; Stopping crit = 2.0551663e-02;
#> Iter 3: Obj fun = -3.1848146; Stopping crit = 2.2051828e-04;
#> Iter 4: Obj fun = -3.1848145; Stopping crit = 3.6574779e-07;
#> Iter 5: Obj fun = -3.1848145; Stopping crit = 1.1800141e-12;
#> Iter 1: Obj fun = -3.1845063; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1778290; Stopping crit = 2.1551027e-02;
#> Iter 3: Obj fun = -3.1777722; Stopping crit = 1.8339738e-04;
#> Iter 4: Obj fun = -3.1777720; Stopping crit = 5.1762284e-07;
#> Iter 5: Obj fun = -3.1777720; Stopping crit = 1.0068629e-11;
#> Iter 1: Obj fun = -3.1874405; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.1808804; Stopping crit = 2.1383706e-02;
#> Iter 3: Obj fun = -3.1807921; Stopping crit = 2.8770685e-04;
#> Iter 4: Obj fun = -3.1807919; Stopping crit = 5.3409422e-07;
#> Iter 5: Obj fun = -3.1807919; Stopping crit = 2.1779687e-12;
#> Iter 1: Obj fun = -3.3017443; Stopping crit = 1.0000000e+00;
#> Iter 2: Obj fun = -3.2954100; Stopping crit = 2.0664261e-02;
#> Iter 3: Obj fun = -3.2953323; Stopping crit = 2.5346890e-04;
#> Iter 4: Obj fun = -3.2953321; Stopping crit = 4.6097119e-07;
#> Iter 5: Obj fun = -3.2953321; Stopping crit = 1.7785887e-12;